Generate pods for all perl modules in lib
Makemaker doesn't provide an easy way to generate pods, which should be
a part of tarball. As of now, using hacky command to run `perldoc`
against all `.pm` files in the lib directory.

Dist:Zilla seems to be a better option as more things are going to work
with simple configuration.